refactor: BasicDecoder + pylint + vulture + black
